refactor: :truck: move provider-config
