+++ apps/taxes/src/modules/taxes/tax-provider.ts
@@ -1,6 +1,6 @@
 import { TaxBaseFragment } from "../../../generated/graphql";
 import { ChannelConfig } from "../channels-configuration/channels-config";
-import { TaxProviderName } from "./providers/config";
+import { TaxProviderName } from "./provider-config";
 import { ResponseTaxPayload } from "./types";
 
 export interface TaxProvider {
